requirements.txt
pygame==2.6.1
pillow==11.2.1
